+++ source/Relativity.Testing.Framework.Api/Strategies/Imaging/Job/NotSupported/ImagingJobRunStrategyNotSupported.cs
@@ -10,12 +10,12 @@ namespace Relativity.Testing.Framework.Api.Strategies
 	{
 		private const string _NOT_SUPPORTED_EXCEPTION_MESSAGE = "The method Run Imaging Job does not support version of Relativity lower than 12.1.";
 
-		public int Run(int workspaceId, int imagingSetId, ImagingSetJobRequest imagingSetJobRequest = null)
+		public long Run(int workspaceId, int imagingSetId, ImagingSetJobRequest imagingSetJobRequest = null)
 		{
 			throw new ArgumentException(_NOT_SUPPORTED_EXCEPTION_MESSAGE);
 		}
 
-		public Task<int> RunAsync(int workspaceId, int imagingSetId, ImagingSetJobRequest imagingSetJobRequest = null)
+		public Task<long> RunAsync(int workspaceId, int imagingSetId, ImagingSetJobRequest imagingSetJobRequest = null)
 		{
 			throw new ArgumentException(_NOT_SUPPORTED_EXCEPTION_MESSAGE);
 		}
